+++ lib/herb/action_view/template_dependencies.rb
@@ -3,6 +3,10 @@
 require "prism"
 require "set"
 
+require_relative "template_dependencies/dependency_collector"
+require_relative "template_dependencies/local_scanner"
+require_relative "template_dependencies/node_dependency_collector"
+
 module Herb
   module ActionView
     class TemplateDependencies
@@ -267,331 +271,6 @@ module Herb
 
       def find_view_root
         @project_path.join("app", "views")
-      end
-    end
-
-    class DependencyCollector < ::Herb::Visitor
-      attr_reader :instance_variables, :constants, :locals_declared,
-                  :locals_received, :helper_calls, :unknown_calls, :render_calls
-
-      def initialize(helper_registry, custom_helpers, prescanned_locals = Set.new)
-        super()
-        @helper_registry = helper_registry
-        @custom_helpers = custom_helpers
-        @instance_variables = Set.new
-        @constants = Set.new
-        @locals_declared = Set.new
-        @locals_received = {} #: Hash[String, String]
-        @helper_calls = Set.new
-        @unknown_calls = Set.new
-        @known_locals = prescanned_locals.dup
-        @render_calls = [] #: Array[Hash[Symbol, untyped]]
-      end
-
-      def visit_erb_node(node)
-        analyze_erb_node(node)
-      end
-
-      def visit_erb_render_node(node)
-        locals = {} #: Hash[String, String]
-
-        node.keywords&.locals&.each do |local|
-          name = local.name&.value
-          value = local.value&.content
-
-          next unless name && value
-
-          value = name if value == "#{name}:"
-          @locals_received[name] = value
-          locals[name] = value
-
-          analyze_ruby_expression(value)
-        end
-
-        if node.static_partial?
-          @render_calls << {
-            partial: node.partial_path,
-            locals: locals,
-            collection: node.keywords&.collection&.value
-          }
-        end
-
-        if node.keywords&.collection
-          analyze_ruby_expression(node.keywords.collection.value)
-        end
-
-        super
-      end
-
-      def visit_erb_strict_locals_node(node)
-        node.locals&.each do |param|
-          @locals_declared.add(param.name&.value) if param.name&.value
-          @known_locals.add(param.name&.value) if param.name&.value
-        end
-
-        super
-      end
-
-      private
-
-      def analyze_erb_node(erb_node)
-        prism_node = erb_node.deserialized_prism_node if erb_node.respond_to?(:deserialized_prism_node)
-
-        if prism_node
-          walk_prism_node(prism_node)
-        else
-          analyze_ruby_expression(erb_node.content&.value&.strip)
-        end
-      rescue StandardError
-        nil
-      end
-
-      def analyze_ruby_expression(code)
-        return if code.nil? || code.empty?
-
-        result = Prism.parse(code)
-        return if result.errors.any?
-
-        walk_prism_node(result.value)
-      rescue StandardError
-        nil
-      end
-
-      def walk_prism_node(node)
-        case node
-        when Prism::InstanceVariableReadNode
-          @instance_variables.add(node.name.to_s)
-        when Prism::LocalVariableWriteNode,
-             Prism::LocalVariableOrWriteNode,
-             Prism::LocalVariableAndWriteNode,
-             Prism::LocalVariableOperatorWriteNode
-          @known_locals.add(node.name.to_s)
-        when Prism::BlockParameterNode, Prism::RequiredParameterNode
-          @known_locals.add(node.name.to_s)
-        when Prism::ConstantReadNode
-          # Standalone constants are just references, not state
-        when Prism::CallNode
-          check_call_node(node)
-        when Prism::LocalVariableReadNode
-          @known_locals.add(node.name.to_s)
-        end
-
-        node.child_nodes.compact.each { |child| walk_prism_node(child) }
-      end
-
-      def check_call_node(node)
-        name = node.name.to_s
-
-        if node.receiver.nil?
-          if @helper_registry.include?(name)
-            @helper_calls.add(name)
-          elsif @custom_helpers.include?(name)
-            @helper_calls.add(name)
-          elsif name == "render"
-            # render calls are handled by visit_erb_render_node
-          elsif !@known_locals.include?(name) && !@locals_received.key?(name) && !@locals_declared.include?(name)
-            @unknown_calls.add(name)
-          end
-        elsif node.receiver.is_a?(Prism::ConstantReadNode)
-          @constants.add("#{node.receiver.name}.#{name}")
-        end
-      end
-    end
-
-    class NodeDependencyCollector < ::Herb::Visitor
-      attr_reader :affected
-
-      def initialize(state, helper_registry, custom_helpers)
-        super()
-        @state = state
-        @helper_registry = helper_registry
-        @custom_helpers = custom_helpers
-        @affected = [] #: Array[Hash[Symbol, untyped]]
-        @path = [] #: Array[Integer]
-        @child_index = [] #: Array[Integer]
-      end
-
-      def visit_document_node(node)
-        visit_children_with_paths(node.child_nodes)
-      end
-
-      def visit_html_element_node(node)
-        visit_children_with_paths(node.child_nodes)
-      end
-
-      def visit_html_open_tag_node(node)
-        node.child_nodes.each_with_index do |child, i|
-          if child.is_a?(Herb::AST::HTMLAttributeNode)
-            check_attribute(child, @path + [i])
-          end
-        end
-      end
-
-      def visit_erb_content_node(node)
-        check_erb_expression(node, :text_content)
-      end
-
-      def visit_erb_if_node(node)
-        check_block_for_state(node, :conditional)
-      end
-
-      def visit_erb_unless_node(node)
-        check_block_for_state(node, :conditional)
-      end
-
-      def visit_erb_case_node(node)
-        check_block_for_state(node, :conditional)
-      end
-
-      def visit_erb_node(node)
-        check_erb_expression(node, :expression)
-      end
-
-      def visit_erb_render_node(node)
-        check_erb_expression(node, :render)
-      end
-
-      private
-
-      def visit_children_with_paths(children)
-        return unless children
-
-        children.each_with_index do |child, index|
-          @path.push(index)
-          visit(child)
-          @path.pop
-        end
-      end
-
-      def check_block_for_state(node, type)
-        all_content = collect_all_expressions(node)
-
-        if all_content.any? { |code| references_state?(code) }
-          location = node.location
-          condition = node.content&.value&.strip
-
-          @affected << {
-            node_path: @path.dup,
-            type: type,
-            expression: condition,
-            location: location ? "#{location.start.line}:#{location.start.column}" : nil
-          }
-        end
-
-        visit_children_with_paths(node.child_nodes)
-      end
-
-      def collect_all_expressions(node)
-        expressions = [] #: Array[String]
-
-        if node.respond_to?(:content) && node.content
-          value = node.content.respond_to?(:value) ? node.content.value&.strip : nil
-
-          expressions << value if value && !value.empty? # steep:ignore
-        end
-
-        children = node.respond_to?(:child_nodes) ? node.child_nodes.compact : [] # steep:ignore
-        children.each { |child| expressions.concat(collect_all_expressions(child)) }
-
-        expressions
-      end
-
-      def check_erb_expression(node, type)
-        code = node.content&.value&.strip
-        return unless code
-
-        if references_state?(code)
-          location = node.location
-
-          @affected << {
-            node_path: @path.dup,
-            type: type,
-            expression: code,
-            location: location ? "#{location.start.line}:#{location.start.column}" : nil
-          }
-        end
-      end
-
-      def check_attribute(attribute_node, path)
-        attribute_name = nil
-
-        attribute_node.child_nodes.each do |child|
-          if child.is_a?(Herb::AST::HTMLAttributeNameNode)
-            first = child.child_nodes&.first
-
-            attribute_name = if first.respond_to?(:content)
-                               content = first.content # steep:ignore
-                               content.respond_to?(:value) ? content.value : content.to_s
-                             end
-          end
-
-          next unless child.is_a?(Herb::AST::HTMLAttributeValueNode)
-
-          child.child_nodes&.each do |value_child|
-            next unless value_child.respond_to?(:content) && value_child.content # steep:ignore
-
-            content = value_child.content # steep:ignore
-            code = (content.respond_to?(:value) ? content.value : content.to_s).strip
-            next unless code && !code.empty? && references_state?(code)
-
-            location = value_child.location # steep:ignore
-
-            @affected << {
-              node_path: path.dup,
-              type: :attribute_value,
-              attribute: attribute_name,
-              expression: code,
-              location: location ? "#{location.start.line}:#{location.start.column}" : nil
-            }
-          end
-        end
-      end
-
-      def references_state?(code)
-        if @state.start_with?("@")
-          code.include?(@state)
-        elsif @state.include?(".")
-          constant = @state.split(".").first
-          code.include?(constant.to_s)
-        else
-          code.match?(/\b#{Regexp.escape(@state)}\b/)
-        end
-      end
-    end
-
-    class LocalScanner < ::Herb::Visitor
-      attr_reader :locals
-
-      def initialize
-        super
-        @locals = Set.new #: Set[String]
-      end
-
-      def visit_erb_node(node)
-        return unless node.respond_to?(:deserialized_prism_node)
-
-        pn = node.deserialized_prism_node
-        collect_locals(pn) if pn
-      rescue StandardError
-        nil
-      end
-
-      private
-
-      def collect_locals(node)
-        case node
-        when Prism::LocalVariableWriteNode,            # title = ...
-             Prism::LocalVariableOrWriteNode,          # title ||= ...
-             Prism::LocalVariableAndWriteNode,         # title &&= ...
-             Prism::LocalVariableOperatorWriteNode     # count += 1
-          @locals.add(node.name.to_s)
-        when Prism::MultiWriteNode
-          node.lefts.each do |target|
-            @locals.add(target.name.to_s) if target.respond_to?(:name) # steep:ignore
-          end
-        end
-
-        node.child_nodes.compact.each { |child| collect_locals(child) }
       end
     end
   end
